BE: charms-sdk test - #3 -  CLI test restructuring
